feat(pack): Add Pack Builder & Make Pack extend List
Adds a pack builder to the Pack class. While not neccesary in the guidelines this could be used later on if more than one pack is needed at once.
This also changes the Pack object to extend List - as it is simply an extension of this. Some boilerplate should be removed by doing this.

The maximum size of a pack has also been hard-coded to 54 (as per the guidelines)