feat: Initial Card Sorter Program (#1)
* feat: add .gitignore

We do not want to commit .idea files to the repository

* feat(codestyle): commit conventions

Add the commit style guidelines from conventional commits

* style: apply K&R Braces

Apply K&R braces which is personally preferred for working in c#

* feat(card): Implement SuitType

Implement the SuitType enum so that we know what each card is whilst retaining the previous integer system

* feat(sort): Implement basic sort objects

Implements a basic system to map int : sort. The sorts are not yet implemented but provides the needed backbone for the `Pack` object.

* refactor(sort): Do not return a new instance of the Pack

as per the guidelines we will retain the same Pack object for the entire project

* feat(pack): Add Pack Builder & Make Pack extend List

Adds a pack builder to the Pack class. While not neccesary in the guidelines this could be used later on if more than one pack is needed at once.
This also changes the Pack object to extend List - as it is simply an extension of this. Some boilerplate should be removed by doing this.

The maximum size of a pack has also been hard-coded to 54 (as per the guidelines)

* feat(sort): Implement signture shuffleCardPack(int typeOfShuffle)

* feat(pack): Implement static card signatures

* feat(sort): imlpement FisherYates Shuffle

* style(fisher-yates): Fix style of FisherYatesShuffle

* feat(sort): Implement RiffleShuffleSort

* fix(project): Add Compile data for Project

* style(project): Fix all remaining styling issues

* refactor: remove unused imports

* fix: make card values readable

* feat: add ToString for the cards

for easier outputting of the card data

* feat: testing data

* build: add .vs/* to .gitignore

* fix: logic errors in code

* feat(docs): Comment entire project

* fix: fix logic error with the pack being 54 instead of 52 lines

* refactor: simplify testing logic

* fix(sort): Ensure pack size is 52

* feat(lib): install log4net

* feat(log): Implement logging

* feat(project): Add README.md

* fix(gitignore): do not commit binary files

* fix(program): Add missing config files and fix initialisation error

* fix(gitignore): do not commit obj folders